initial implementation of CLI tool
To make this work, I also stopped creating a Blog instance at all
because we are focused on Posts. I think I may, in the future, switch
this to be Sentence oriented, although I don't know whether that will
really be the change I want to see in the app...

Anyway, this commit also includes a restructure, although I don't really
know how this is *supposed* to work, I'm mimicking the structure of
other applications. Hence, you can only run `bin/omnivore` from the root
of the project. I will probably put a small amount of effort into fixing
this in the future.